formatting

+++ utils/weigh_chars.c
@@ -1,4 +1,4 @@
-void weigh_chars(char c, int *scale, char weight1, char weight2)
+void weigh_chars(char c, short* scale, char weight1, char weight2)
 {
     if (c == weight1) {
         (*scale)++;
